+++ src/FreeElephants/Configuration/Console/Exception/RuntimeException.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace FreeElephants\Configuration\Console\Exception;
+
+use FreeElephants\Configuration\Exception\ConfigurationExceptionInterface;
+
+/**
+ *
+ * @author samizdam
+ *
+ */
+class RuntimeException extends \Symfony\Component\Console\Exception\RuntimeException implements ConfigurationExceptionInterface
+{
+}