+++ src/FreeElephants/Configuration/Writer/PhpWriter.php
@@ -2,11 +2,22 @@
 
 namespace FreeElephants\Configuration\Writer;
 
+use FreeElephants\BitMask\BitMaskTrait;
 /**
+ *
  * @author samizdam
  */
 class PhpWriter implements WriterInterface
 {
+
+    use BitMaskTrait;
+
+    const OPTION_ADD_PHP_OPEN_TAG = 0b0001;
+
+    const OPTION_ADD_RETURN_STATEMENT = 0b0010;
+
+    private $options = self::OPTION_ADD_RETURN_STATEMENT;
+
     public function writeFile($filename, $data)
     {
         file_put_contents($filename, $this->toString($data));
@@ -14,6 +25,20 @@ class PhpWriter implements WriterInterface
 
     public function toString($data)
     {
-        return var_export($data, true);
+        $string = var_export($data, true);
+        if ($this->hasFlag(self::OPTION_ADD_PHP_OPEN_TAG)) {
+            $string = "<php\n" . $string;
+        }
+        if ($this->hasFlag(self::OPTION_ADD_RETURN_STATEMENT)) {
+            $string = "return\n" . $string;
+        }
+        return $string;
     }
+
+
+    protected function getBitFieldValue()
+    {
+        return $this->options;
+    }
+
 }