+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog 
 
+## Unrealised
+## Added
+- Console converter for different formats.  
+
 ## 0.9.0
 ### Added 
 - Yaml, Json and native php-arrays support.